restructure project
