+++ tc-default_themes/Cargo.toml
@@ -1,9 +1,12 @@
+# TODO: When the loaders are implemented this crate will be deleted
+# -> the default_themes will move into the config_loader crate
 [package]
 name = "tc-default_themes"
 edition = "2024"
 
 [dependencies]
 tc-user_config_loader = { path = "../tc-user_config_loader/" }
+tc-models             = { path = "../tc-models/" }
 
 # External dependencies
 ratatui               = { workspace = true }
