+++ tc-models/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "tc-models"
+edition = "2024"
+
+[dependencies]
+ratatui = { workspace = true }
+
+# External dependencies
+unicode-segmentation = { workspace = true }
+chrono               = { workspace = true }
