+++ tc-models/src/clock.rs
@@ -1,0 +1,5 @@
+use ratatui::widgets::Paragraph;
+
+pub trait Clock {
+    fn draw_clockface(&self, clock_format: &str) -> (Paragraph, usize, usize);
+}
