+++ tc-models/src/color_clock.rs
@@ -1,5 +1,5 @@
-use crate::clock_types::helper::{art_block, combine_ascii_art_while_applying_led};
-use crate::models::clock::Clock;
+use crate::clock::Clock;
+use crate::helper::{art_block, combine_ascii_art_while_applying_led};
 use chrono::{Local, Timelike};
 use ratatui::{style::Color, widgets::Paragraph};
