+++ tc-models/src/digital_clock.rs
@@ -1,4 +1,4 @@
-use crate::models::clock::Clock;
+use crate::clock::Clock;
 use ratatui::widgets::Paragraph;
 
 pub struct DigitalClock {
