+++ tc-models/src/lib.rs
@@ -1,4 +1,8 @@
 pub mod analog_clock;
+pub mod clock;
 pub mod color_clock;
+pub mod colorscheme;
 pub mod digital_clock;
 pub mod helper;
+pub mod quote;
+
