+++ tc-models/src/quote.rs
