+++ tc-user_config_loader/Cargo.toml
@@ -3,8 +3,10 @@ name = "tc-user_config_loader"
 edition = "2024"
 
 [dependencies]
+tc-models            = { path = "../tc-models/" }
+
+# External dependencies
 ratatui              = { workspace = true }
 chrono               = { workspace = true }
-unicode-segmentation = { workspace = true }
 toml                 = { workspace = true }
 serde                = { workspace = true }
