--- tc-user_config_loader/src/loader/mod.rs
@@ -1,4 +1,0 @@
-pub mod clock_face_loader;
-pub mod colorscheme_loader;
-mod load_error;
-pub mod quote_loader;
