--- tc-user_config_loader/src/models/clock.rs
@@ -1,25 +1,0 @@
-use crate::{
-    clock_types::color_clock::ColorClock,
-    loader::clock_face_loader::{AnalogClockConfig, ColorClockConfig, DigitalClockConfig},
-};
-use ratatui::style::Color;
-use ratatui::widgets::Paragraph;
-use std::str::FromStr;
-
-pub trait Clock {
-    fn draw_clockface(&self, clock_format: &str) -> (Paragraph, usize, usize);
-}
-
-impl From<ColorClockConfig> for ColorClock {
-    fn from(config: ColorClockConfig) -> ColorClock {
-        ColorClock::new(
-            config.hour,
-            config.minutes,
-            config.seconds,
-            config.hour_coords,
-            config.minute_coords,
-            config.second_coords,
-            Color::from_str(config.accent_color.as_str()).unwrap_or(Color::White),
-        )
-    }
-}
