--- tc-user_config_loader/src/models/mod.rs
@@ -1,3 +1,0 @@
-pub mod clock;
-pub mod colorscheme;
-pub mod quote;
